Vad vi gjort än så länge
inte mycket lmao

but who is lmao ? chinese hacker?
lmfao? is that his sister?
lmbfao?
we now know that lmao, is working with lmfao and lmbfao
